added _components directory
