+++ task1/src/App.css
@@ -72,9 +72,9 @@
   display: block;
 }
 
-.btn-menu {
+/* .btn-menu {
 
-}
+} */
 
 /* .text-center {} */
