+++ task1/src/_components/MenuItem.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './App.css';
+import '../App.css'; 
 
 const MenuItem = ({img, title, link}) => {
     return (
